requirements.txt
# Core
pandas>=1.5.0,<2.0.0
numpy>=1.23.0,<1.25.0
yfinance>=0.2.18

# Technical Analysis - ta-lib alternative using pandas
ta>=0.10.2

# Sentiment Analysis
vaderSentiment>=3.3.2

# News Collection
feedparser>=6.0.10
requests>=2.28.0
beautifulsoup4>=4.11.0

# CLI & Display
rich>=13.3.0
click>=8.1.0

# Configuration
python-dotenv>=0.21.0
pyyaml>=6.0

# Testing
pytest>=7.2.0